MainMenuPanel is IGameMenuPanel, move menu into its own namespace
