+++ .addon
@@ -7,8 +7,8 @@
   "Schema": 1,
   "HasAssets": true,
   "AssetsPath": "",
-  "Resources": "ui/*\nfonts/*\ndata/*",
-  "MenuResources": null,
+  "Resources": "ui/*\nfonts/*\ndata/*\n.localization/*",
+  "MenuResources": "/UI/Menu/*\n/Sounds/UI/*\n/.localization/*",
   "HasCode": true,
   "CodePath": "code",
   "PackageReferences": [
