+++ code/Game.cs
@@ -3,6 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using Amper.FPS;
 using TFS2.UI;
+using TFS2.Menu;
 
 namespace TFS2;
