+++ code/UI/HUD/Server/ClassSelection.cs
@@ -3,6 +3,7 @@ using Sandbox.UI;
 using Sandbox.UI.Construct;
 using System;
 using System.Linq;
+using TFS2.Menu;
 
 namespace TFS2.UI;
