+++ code/UI/HUD/Server/ServerMessage.cs
@@ -1,6 +1,7 @@
 ﻿using Sandbox;
 using Sandbox.UI;
 using Sandbox.UI.Construct;
+using TFS2.Menu;
 
 namespace TFS2.UI;
