+++ code/UI/HUD/Server/TeamSelection.cs
@@ -2,6 +2,7 @@
 using Sandbox.UI;
 using Sandbox.UI.Construct;
 using System.Linq;
+using TFS2.Menu;
 
 namespace TFS2.UI;
