+++ code/UI/HUD/Server/TeamSelection.razor
@@ -1,4 +1,5 @@
 ﻿@using Sandbox.UI;
+@using TFS2.Menu;
 
 @attribute [StyleSheet("/UI/HUD/Server/TeamSelection.scss")]
