+++ code/UI/Menu/BlogView.razor
@@ -1,6 +1,6 @@
 ﻿@using System;
 @using Sandbox.UI;
-@namespace TFS2.UI
+@namespace TFS2.Menu
 @attribute [StyleSheet("UI/Menu/BlogView.scss")]
 
 @inherits MenuOverlay
