+++ code/UI/Menu/Items/ClassLoadout.razor
@@ -3,7 +3,7 @@
 
 @attribute [StyleSheet("/UI/Menu/Items/ClassLoadout.scss")]
 
-@namespace TFS2.UI
+@namespace TFS2.Menu
 @inherits MenuOverlay
 
 <root>
