+++ code/UI/Menu/Items/ItemPicker.razor
@@ -3,7 +3,7 @@
 
 @attribute [StyleSheet("/UI/Menu/Items/ItemPicker.scss")]
 
-@namespace TFS2.UI
+@namespace TFS2.Menu
 
 <root>
     @if (PlayerClass == null)
