+++ code/UI/Menu/Items/ItemSelection.razor
@@ -3,7 +3,7 @@
 
 @attribute [StyleSheet("/UI/Menu/Items/ItemSelection.scss")]
 
-@namespace TFS2.UI
+@namespace TFS2.Menu
 @inherits MenuOverlay
 
 <root>
