+++ code/UI/Menu/Items/ItemsPage.razor
@@ -2,7 +2,7 @@
 
 @attribute [StyleSheet("/UI/Menu/Items/ItemPage.scss")]
 
-@namespace TFS2.UI
+@namespace TFS2.Menu
 @inherits MenuOverlay
 
 <root>
