+++ code/UI/Menu/MainMenuPanel.scss
@@ -10,15 +10,6 @@ MainMenuPanel {
     flex-direction: column;
     backdrop-filter: blur(5px);
     background-color: #100a0194;
-    opacity: 0;
-    transform: scale(1.01);
-    transition: all .05s ease;
-
-    &.visible {
-        opacity: 1;
-        pointer-events: all;
-        transform: scale(1);
-    }
 
     .container {
         padding: 0 50px;
