+++ code/UI/Menu/QuitDialog.cs
@@ -1,7 +1,7 @@
 ﻿using Sandbox;
 using Sandbox.UI;
 
-namespace TFS2.UI;
+namespace TFS2.Menu;
 
 /// <summary>
 /// Popup dialog box asking the player if they want to quit i.e. MessagePanel.cs
