+++ code/UI/Menu/QuitDialog.razor
@@ -2,7 +2,7 @@
 
 @attribute [StyleSheet("/UI/Menu/QuitDialog.scss")]
 
-@namespace TFS2.UI
+@namespace TFS2.Menu
 @inherits MenuOverlay
 
 <root>
