+++ code/UI/Menu/Settings/SettingsMenu.razor
@@ -3,7 +3,7 @@
 @using Amper.FPS;
 @using Sandbox.UI;
 
-@namespace TFS2
+@namespace TFS2.Menu
 @attribute [StyleSheet("/ui/Menu/Settings/SettingsMenu.scss")]
 @inherits MenuOverlay
