+++ code/UI/TFRootPanel.razor
@@ -8,7 +8,9 @@
 @inherits RootPanel
 <root>
     <GameHUD></GameHUD>
-    <MainMenuPanel></MainMenuPanel>
+
+    <!-- Developer -->
+    <DevMenu></DevMenu>
 	<TauntMenu></TauntMenu>
     <!--
         UNUSED
